Load restaurants in user's order
On opening the page, it goes to load menu for all restaurants.
Right now it's doing that in original embedded order.

But as user may have changed the order (and kept some of the menus open),
there may be restaurants from the end of the list visible on top of the
page, but being loaded as last (staying empty for few seconds).

This changes the loading to be done in same order as user configured it,
and so first showing it for those restaurants which user cares about most.